requirements.txt
torch==1.10.0
torchvision==0.11.1
matplotlib==3.4.3
Pillow==8.4.0
einops==0.3.2
wandb==0.12.1
torch-distributed==1.10.1
